office laptop changes

+++ app/Services/FileUploadService.php
@@ -12,17 +12,19 @@ class FileUploadService
         // Verify that the image belongs to the correct session
         $tempImagePath = session('temp'); 
 
-        if (Storage::disk('public')->exists($tempImagePath)) {   
-            // Add the temporarily stored image to the media collection
-            $model->addMedia(storage_path("app/public/{$tempImagePath}"))->toMediaCollection('default');
-
-            // // Clean up temporary storage (if needed)
-            Storage::disk('public')->deleteDirectory('temp');
-
-            session()->forget('temp');
-        } else {
-            // Image doesn't belong to the correct session
-            // Handle accordingly (e.g., show an error, log the incident)
+        if($tempImagePath){ 
+            if (Storage::disk('public')->exists($tempImagePath)) {   
+                // Add the temporarily stored image to the media collection
+                $model->addMedia(storage_path("app/public/{$tempImagePath}"))->toMediaCollection('default');
+    
+                // // Clean up temporary storage (if needed)
+                Storage::disk('public')->deleteDirectory('temp');
+    
+                session()->forget('temp');
+            } else {
+                // Image doesn't belong to the correct session
+                // Handle accordingly (e.g., show an error, log the incident)
+            }
         }
     }
 }
